Changed TransportFare's Datatype to Float

+++ src/main/java/com/rfid/esptocloud/domain/Card.java
@@ -26,6 +26,6 @@ public class Card {
     @Column("new_balance")
     public Float newBalance;
     @Column("transport_fare")
-    public Integer transportFare;
+    public Float transportFare;
 
 }
